# docs/api/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "API Documentation",
    "version" : "1.0"
  },
  "paths" : {
    "/webauthn/v1/register/challenge/{username}" : {
      "get" : {
        "summary" : "处理注册挑战请求。",
        "description" : "处理注册挑战请求。",
        "parameters" : [ {
          "name" : "username",
          "in" : "path",
          "description" : "用户名",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "PublicKeyCredentialCreationOptions对象",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "rp" : {
                      "type" : "object",
                      "properties" : {
                        "name" : {
                          "type" : "string",
                          "description" : ""
                        },
                        "id" : {
                          "type" : "string",
                          "description" : ""
                        }
                      },
                      "description" : ""
                    },
                    "user" : {
                      "type" : "object",
                      "properties" : {
                        "name" : {
                          "type" : "string",
                          "description" : ""
                        },
                        "id" : {
                          "type" : "array",
                          "description" : "",
                          "items" : {
                            "type" : "integer"
                          }
                        },
                        "displayName" : {
                          "type" : "string",
                          "description" : ""
                        }
                      },
                      "description" : ""
                    },
                    "challenge" : {
                      "type" : "object",
                      "properties" : { },
                      "description" : ""
                    },
                    "pubKeyCredParams" : {
                      "type" : "array",
                      "description" : "",
                      "items" : {
                        "type" : "object",
                        "properties" : {
                          "type" : {
                            "type" : "object",
                            "properties" : {
                              "value" : {
                                "type" : "string",
                                "description" : ""
                              }
                            },
                            "description" : ""
                          },
                          "alg" : {
                            "type" : "object",
                            "properties" : {
                              "value" : {
                                "type" : "integer",
                                "description" : "",
                                "format" : "int64"
                              }
                            },
                            "description" : ""
                          }
                        },
                        "description" : "com.webauthn4j.data.PublicKeyCredentialParameters"
                      }
                    },
                    "timeout" : {
                      "type" : "integer",
                      "description" : "",
                      "format" : "int64"
                    },
                    "excludeCredentials" : {
                      "type" : "array",
                      "description" : "",
                      "items" : {
                        "type" : "object",
                        "properties" : {
                          "type" : {
                            "type" : "object",
                            "properties" : {
                              "value" : {
                                "type" : "string",
                                "description" : ""
                              }
                            },
                            "description" : ""
                          },
                          "id" : {
                            "type" : "array",
                            "description" : "",
                            "items" : {
                              "type" : "integer"
                            }
                          },
                          "transports" : {
                            "type" : "array",
                            "description" : "",
                            "items" : {
                              "type" : "object",
                              "properties" : {
                                "value" : {
                                  "type" : "string",
                                  "description" : ""
                                }
                              },
                              "description" : "com.webauthn4j.data.AuthenticatorTransport"
                            }
                          }
                        },
                        "description" : "com.webauthn4j.data.PublicKeyCredentialDescriptor"
                      }
                    },
                    "authenticatorSelection" : {
                      "type" : "object",
                      "properties" : {
                        "authenticatorAttachment" : {
                          "type" : "object",
                          "properties" : {
                            "value" : {
                              "type" : "string",
                              "description" : ""
                            }
                          },
                          "description" : ""
                        },
                        "requireResidentKey" : {
                          "type" : "boolean",
                          "description" : ""
                        },
                        "residentKey" : {
                          "type" : "object",
                          "properties" : {
                            "value" : {
                              "type" : "string",
                              "description" : ""
                            }
                          },
                          "description" : ""
                        },
                        "userVerification" : {
                          "type" : "object",
                          "properties" : {
                            "value" : {
                              "type" : "string",
                              "description" : ""
                            }
                          },
                          "description" : ""
                        }
                      },
                      "description" : ""
                    },
                    "attestation" : {
                      "type" : "object",
                      "properties" : {
                        "value" : {
                          "type" : "string",
                          "description" : ""
                        }
                      },
                      "description" : ""
                    },
                    "extensions" : {
                      "type" : "object",
                      "properties" : {
                        "appid" : {
                          "type" : "string",
                          "description" : ""
                        },
                        "appidExclude" : {
                          "type" : "string",
                          "description" : ""
                        },
                        "uvm" : {
                          "type" : "boolean",
                          "description" : ""
                        },
                        "credProps" : {
                          "type" : "boolean",
                          "description" : ""
                        },
                        "credentialProtectionPolicy" : {
                          "type" : "string",
                          "description" : "{USER_VERIFICATION_OPTIONAL=USER_VERIFICATION_OPTIONAL, USER_VERIFICATION_OPTIONAL_WITH_CREDENTIAL_ID_LIST=USER_VERIFICATION_OPTIONAL_WITH_CREDENTIAL_ID_LIST, USER_VERIFICATION_REQUIRED=USER_VERIFICATION_REQUIRED}",
                          "enum" : [ "USER_VERIFICATION_OPTIONAL", "USER_VERIFICATION_OPTIONAL_WITH_CREDENTIAL_ID_LIST", "USER_VERIFICATION_REQUIRED" ]
                        },
                        "enforceCredentialProtectionPolicy" : {
                          "type" : "boolean",
                          "description" : ""
                        },
                        "hmacCreateSecret" : {
                          "type" : "boolean",
                          "description" : ""
                        },
                        "hmacGetSecret" : {
                          "type" : "object",
                          "properties" : {
                            "salt1" : {
                              "type" : "array",
                              "description" : "",
                              "items" : {
                                "type" : "integer"
                              }
                            },
                            "salt2" : {
                              "type" : "array",
                              "description" : "",
                              "items" : {
                                "type" : "integer"
                              }
                            }
                          },
                          "description" : ""
                        }
                      },
                      "description" : ""
                    }
                  },
                  "description" : "PublicKeyCredentialCreationOptions对象"
                }
              }
            }
          }
        }
      }
    },
    "/webauthn/v1/register/credential" : {
      "post" : {
        "summary" : "处理注册凭据请求。",
        "description" : "处理注册凭据请求。",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "id" : {
                    "type" : "string",
                    "description" : ""
                  },
                  "rawId" : {
                    "type" : "string",
                    "description" : ""
                  },
                  "response" : {
                    "type" : "object",
                    "properties" : {
                      "attestationObject" : {
                        "type" : "string",
                        "description" : ""
                      },
                      "clientDataJSON" : {
                        "type" : "string",
                        "description" : ""
                      }
                    },
                    "description" : ""
                  },
                  "type" : {
                    "type" : "string",
                    "description" : ""
                  }
                },
                "description" : "CredentialRequest对象"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Map对象",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "credentialId" : {
                      "type" : "string"
                    }
                  },
                  "description" : "Map对象"
                }
              }
            }
          }
        }
      }
    },
    "/webauthn/v1/login/challenge/{username}" : {
      "get" : {
        "summary" : "处理登录挑战请求。",
        "description" : "处理登录挑战请求。",
        "parameters" : [ {
          "name" : "username",
          "in" : "path",
          "description" : "用户名",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "PublicKeyCredentialRequestOptions对象",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "challenge" : {
                      "type" : "object",
                      "properties" : { },
                      "description" : ""
                    },
                    "timeout" : {
                      "type" : "integer",
                      "description" : "",
                      "format" : "int64"
                    },
                    "rpId" : {
                      "type" : "string",
                      "description" : ""
                    },
                    "allowCredentials" : {
                      "type" : "array",
                      "description" : "",
                      "items" : {
                        "type" : "object",
                        "properties" : {
                          "type" : {
                            "type" : "object",
                            "properties" : {
                              "value" : {
                                "type" : "string",
                                "description" : ""
                              }
                            },
                            "description" : ""
                          },
                          "id" : {
                            "type" : "array",
                            "description" : "",
                            "items" : {
                              "type" : "integer"
                            }
                          },
                          "transports" : {
                            "type" : "array",
                            "description" : "",
                            "items" : {
                              "type" : "object",
                              "properties" : {
                                "value" : {
                                  "type" : "string",
                                  "description" : ""
                                }
                              },
                              "description" : "com.webauthn4j.data.AuthenticatorTransport"
                            }
                          }
                        },
                        "description" : "com.webauthn4j.data.PublicKeyCredentialDescriptor"
                      }
                    },
                    "userVerification" : {
                      "type" : "object",
                      "properties" : {
                        "value" : {
                          "type" : "string",
                          "description" : ""
                        }
                      },
                      "description" : ""
                    },
                    "extensions" : {
                      "type" : "object",
                      "properties" : {
                        "appid" : {
                          "type" : "string",
                          "description" : ""
                        },
                        "appidExclude" : {
                          "type" : "string",
                          "description" : ""
                        },
                        "uvm" : {
                          "type" : "boolean",
                          "description" : ""
                        },
                        "credProps" : {
                          "type" : "boolean",
                          "description" : ""
                        },
                        "credentialProtectionPolicy" : {
                          "type" : "string",
                          "description" : "{USER_VERIFICATION_OPTIONAL=USER_VERIFICATION_OPTIONAL, USER_VERIFICATION_OPTIONAL_WITH_CREDENTIAL_ID_LIST=USER_VERIFICATION_OPTIONAL_WITH_CREDENTIAL_ID_LIST, USER_VERIFICATION_REQUIRED=USER_VERIFICATION_REQUIRED}",
                          "enum" : [ "USER_VERIFICATION_OPTIONAL", "USER_VERIFICATION_OPTIONAL_WITH_CREDENTIAL_ID_LIST", "USER_VERIFICATION_REQUIRED" ]
                        },
                        "enforceCredentialProtectionPolicy" : {
                          "type" : "boolean",
                          "description" : ""
                        },
                        "hmacCreateSecret" : {
                          "type" : "boolean",
                          "description" : ""
                        },
                        "hmacGetSecret" : {
                          "type" : "object",
                          "properties" : {
                            "salt1" : {
                              "type" : "array",
                              "description" : "",
                              "items" : {
                                "type" : "integer"
                              }
                            },
                            "salt2" : {
                              "type" : "array",
                              "description" : "",
                              "items" : {
                                "type" : "integer"
                              }
                            }
                          },
                          "description" : ""
                        }
                      },
                      "description" : ""
                    }
                  },
                  "description" : "PublicKeyCredentialRequestOptions对象"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}